% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coordinates.R
\name{apply_symmetry_operations}
\alias{apply_symmetry_operations}
\title{Apply Symmetry Operations to Generate a Full Unit Cell}
\usage{
apply_symmetry_operations(
  atomic_coordinates,
  symmetry_operations,
  unit_cell_metrics,
  tolerance = 1e-04
)
}
\arguments{
\item{atomic_coordinates}{A \code{data.table} of asymmetric atoms.}

\item{symmetry_operations}{A \code{data.table} of operations.}

\item{unit_cell_metrics}{A \code{data.table} of cell parameters (needed for distance merging).}

\item{tolerance}{Numeric. The distance tolerance in Angstroms for merging close atoms.
Default is 1e-4.}
}
\value{
A \code{data.table} containing all unique atomic positions in the unit cell.
}
\description{
Generates all symmetry-equivalent atomic positions within the unit cell.
}
\examples{
ac <- data.table::data.table(Label = "A", x_a = 0, y_b = 0, z_c = 0)
ops <- data.table::data.table(x = c("x", "-x"), y = c("y", "-y"), z = c("z", "-z"))
uc <- data.table::data.table(`_cell_length_a` = 10, `_cell_length_b` = 10,
                             `_cell_length_c` = 10, `_cell_angle_alpha` = 90,
                             `_cell_angle_beta` = 90, `_cell_angle_gamma` = 90)
apply_symmetry_operations(ac, ops, uc)
}
\seealso{
Other coordinate processors: 
\code{\link{calculate_expansion_factors}()},
\code{\link{expand_transformed_coords}()}
}
\concept{coordinate processors}
