% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{crystal_nn}
\alias{crystal_nn}
\title{Identify Atomic Bonds using CrystalNN}
\usage{
crystal_nn(
  distances,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics,
  cutoff_length = 7,
  x_diff_weight = 3,
  porous_adjustment = TRUE,
  distance_cutoffs = c(0.5, 1),
  cation_anion = FALSE,
  weighted_cn = FALSE
)
}
\arguments{
\item{distances}{Ignored (CrystalNN generates its own Voronoi basis).}

\item{atomic_coordinates}{Primary atom set.}

\item{expanded_coords}{Expanded supercell.}

\item{unit_cell_metrics}{Cell parameters.}

\item{cutoff_length}{Numeric. Cutoff in Angstroms for initial neighbor search
(default 7.0).}

\item{x_diff_weight}{Numeric. Electronegativity difference weight (default 3.0).}

\item{porous_adjustment}{Logical. If TRUE, adjusts Voronoi weights to
better describe layered / porous structures (default TRUE).}

\item{distance_cutoffs}{Numeric vector. Penalizes neighbor distances greater than
sum of covalent radii plus these cutoffs. Set to NULL to disable.}

\item{cation_anion}{Logical. If TRUE, restricts bonding targets to sites
with opposite or zero charge (requires oxidation states).}

\item{weighted_cn}{Logical. If FALSE (default), returns neighbors for the
most probable coordination environment with weight 1.0. If TRUE, returns all
potential neighbors with fractional probabilities.}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Poer of Pymatgen's \code{CrystalNN} algorithm from the ground up.
It uses a Voronoi-based algorithm and solid angle weights to determine the
probability of various coordination environments. It modifies probability using
smooth distance cutoffs and Pauling electronegativity differences. The output
is either the most probable coordination environment (\code{weighted_cn = FALSE})
or a weighted list of coordination environments (\code{weighted_cn = TRUE}).
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{econ_nn}()},
\code{\link{minimum_distance}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
