% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crystract-package.R
\docType{package}
\name{crystract-package}
\alias{crystract}
\alias{crystract-package}
\title{crystract: Tools to Extract Data from CIF Files for Crystallography}
\description{
Provides a suite of functions to parse Crystallographic Information
Files (.cif), extracting essential data such as chemical formulas, unit cell
parameters, atomic coordinates, and symmetry operations. It also includes
tools to calculate interatomic distances, identify bonded pairs using various
algorithms (Minimum Distance, Brunner's, EconNN, CrystalNN, Voronoi), determine nearest
neighbor counts, and calculate bond angles. The package is designed to
facilitate the preparation of crystallographic data for further analysis,
including machine learning applications in materials science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://prabhulab.github.io/ml-crystals/}
}

}
\keyword{internal}
