% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{filter_by_elements}
\alias{filter_by_elements}
\title{Filter Distances by Element Symbols}
\usage{
filter_by_elements(distances, atomic_coordinates, elements_to_exclude)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances.}

\item{atomic_coordinates}{A \code{data.table} of atomic coordinates.}

\item{elements_to_exclude}{A character vector of element symbols to exclude.}
}
\value{
A \code{data.table} of distances with the specified elements removed.
}
\description{
Removes any distance pair where at least one of the atoms
corresponds to a specified list of element symbols.
}
\seealso{
Other property calculators: 
\code{\link{calculate_angles}()},
\code{\link{calculate_distances}()},
\code{\link{calculate_neighbor_counts}()},
\code{\link{calculate_weighted_neighbor_counts}()},
\code{\link{filter_atoms_by_symbol}()},
\code{\link{filter_by_wyckoff_symbol}()}
}
\concept{property calculators}
