% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{voronoi_nn}
\alias{voronoi_nn}
\title{Identify Atomic Bonds using Voronoi Tessellation}
\usage{
voronoi_nn(
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics,
  cutoff = 13,
  tol = 0
)
}
\arguments{
\item{atomic_coordinates}{A \code{data.table} of the primary (asymmetric) atom set.}

\item{expanded_coords}{A \code{data.table} of atoms in the expanded supercell.}

\item{unit_cell_metrics}{A \code{data.table} with cell parameters.}

\item{cutoff}{Distance cutoff (default 13.0).}

\item{tol}{Tolerance for solid angle weights (default 0).}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Performs 3D Voronoi analysis on the supercell.
strictly mapping Asymmetric Unit atoms to their Supercell equivalents for processing.
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{crystal_nn}()},
\code{\link{econ_nn}()},
\code{\link{minimum_distance}()}
}
\concept{bonding algorithms}
