% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_coordinates.R
\name{apply_symmetry_operations}
\alias{apply_symmetry_operations}
\title{Apply Symmetry Operations to Generate a Full Unit Cell}
\usage{
apply_symmetry_operations(atomic_coordinates, symmetry_operations)
}
\arguments{
\item{atomic_coordinates}{A \code{data.table} of asymmetric atoms from
\code{extract_atomic_coordinates}.}

\item{symmetry_operations}{A \code{data.table} of operations from
\code{extract_symmetry_operations}.}
}
\value{
A \code{data.table} containing all unique atomic positions in the unit cell.
}
\description{
Generates all symmetry-equivalent atomic positions within the unit cell
by applying the list of symmetry operations to the initial set of
asymmetric atoms.
}
\examples{
cif_file <- system.file("extdata", "ICSD422.cif", package = "crysmal")
if (file.exists(cif_file)) {
  cif_content <- read_cif_files(cif_file)[[1]]
  atoms <- extract_atomic_coordinates(cif_content)
  sym_ops <- extract_symmetry_operations(cif_content)
  full_cell_coords <- apply_symmetry_operations(atoms, sym_ops)
  print(full_cell_coords)
}
}
\seealso{
Other coordinate processors: 
\code{\link{expand_transformed_coords}()}
}
\concept{coordinate processors}
