% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{brunner_nn_reciprocal}
\alias{brunner_nn_reciprocal}
\title{Identify Atomic Bonds using Brunner's Method (Reciprocal)}
\usage{
brunner_nn_reciprocal(distances, tol = 0)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances.}

\item{tol}{A small distance offset (default 0.0) to add to the cutoff.}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
An alternative bonding detection method using the
largest reciprocal gap. Matches \code{pymatgen.analysis.local_env.BrunnerNNReciprocal}.
}
\seealso{
Other bonding algorithms: 
\code{\link{crystal_nn}()},
\code{\link{econ_nn}()},
\code{\link{minimum_distance}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
