% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{calculate_angles}
\alias{calculate_angles}
\title{Calculate Bond Angles}
\usage{
calculate_angles(
  bonded_pairs,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics
)
}
\arguments{
\item{bonded_pairs}{Data.table of bonded atoms.}

\item{atomic_coordinates}{Data.table of asymmetric atom coordinates.}

\item{expanded_coords}{Data.table of supercell atom coordinates.}

\item{unit_cell_metrics}{Data.table with unit cell parameters.}
}
\value{
A \code{data.table} of all unique bond angles.
}
\description{
Calculates all bond angles centered on each atom, formed by pairs
of its bonded neighbors.
}
\seealso{
Other property calculators: 
\code{\link{calculate_distances}()},
\code{\link{calculate_neighbor_counts}()},
\code{\link{calculate_weighted_average_network_distance}()},
\code{\link{filter_atoms_by_symbol}()},
\code{\link{filter_by_elements}()},
\code{\link{filter_by_wyckoff_symbol}()},
\code{\link{filter_ghost_distances}()}
}
\concept{property calculators}
