% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{calculate_neighbor_counts}
\alias{calculate_neighbor_counts}
\title{Calculate Coordination Numbers}
\usage{
calculate_neighbor_counts(bonded_pairs_table)
}
\arguments{
\item{bonded_pairs_table}{A \code{data.table} of bonded pairs.}
}
\value{
A \code{data.table} with columns 'Atom' and 'CoordinationNumber'.
}
\description{
Counts the number of nearest neighbors for each central atom based on
a table of bonded pairs.
}
\seealso{
Other property calculators: 
\code{\link{calculate_angles}()},
\code{\link{calculate_distances}()},
\code{\link{calculate_weighted_neighbor_counts}()},
\code{\link{filter_atoms_by_symbol}()},
\code{\link{filter_by_elements}()},
\code{\link{filter_by_wyckoff_symbol}()}
}
\concept{property calculators}
