% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{calculate_weighted_average_network_distance}
\alias{calculate_weighted_average_network_distance}
\title{Calculate Weighted Average Network Distance}
\usage{
calculate_weighted_average_network_distance(
  distances,
  atomic_coordinates,
  wyckoff_symbols
)
}
\arguments{
\item{distances}{A \code{data.table} of cleaned interatomic distances.}

\item{atomic_coordinates}{A \code{data.table} from \code{extract_atomic_coordinates}.}

\item{wyckoff_symbols}{A character vector of Wyckoff symbols to include.}
}
\value{
A single numeric value for the weighted average distance.
}
\description{
Calculates a weighted average distance from a pre-filtered table.
}
\seealso{
Other property calculators: 
\code{\link{calculate_angles}()},
\code{\link{calculate_distances}()},
\code{\link{calculate_neighbor_counts}()},
\code{\link{filter_atoms_by_symbol}()},
\code{\link{filter_by_elements}()},
\code{\link{filter_by_wyckoff_symbol}()},
\code{\link{filter_ghost_distances}()}
}
\concept{property calculators}
