% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covalent_radii}
\alias{covalent_radii}
\title{Atomic Radii Data for Bond-Length Estimation}
\format{
A data table with two columns:
\describe{
\item{Symbol}{The chemical symbol of the element.}
\item{Radius}{The atomic radius in Angstroms.}
}
}
\source{
J. Emsley. The Elements. Third edition 1998, Oxford University Press.
As provided by Julia-Maria Huebner.
}
\usage{
covalent_radii
}
\description{
A data.table containing atomic radii for elements. This data is used for
estimating plausible bond lengths to filter out non-physical "ghost" distances
that can occur in disordered structures. Radii are in Angstroms (Å).
}
\keyword{datasets}
