% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{crystal_nn}
\alias{crystal_nn}
\title{Identify Atomic Bonds using CrystalNN}
\usage{
crystal_nn(
  distances,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics,
  cutoff_length = 13,
  x_diff_weight = 3,
  porosity_adjustment = TRUE,
  distance_cutoffs = c(0.5, 1)
)
}
\arguments{
\item{distances}{Ignored.}

\item{atomic_coordinates}{Primary atom set.}

\item{expanded_coords}{Expanded supercell.}

\item{unit_cell_metrics}{Cell parameters.}

\item{cutoff_length}{Voronoi cutoff.}

\item{x_diff_weight}{Electronegativity weight (default 3.0).}

\item{porosity_adjustment}{Logical (default TRUE).}

\item{distance_cutoffs}{Vector c(0.5, 1.0).}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Port of Pymatgen's \code{CrystalNN} weighting logic.
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{econ_nn}()},
\code{\link{minimum_distance}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
