% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{crystal_nn}
\alias{crystal_nn}
\title{Identify Atomic Bonds using CrystalNN}
\usage{
crystal_nn(
  distances,
  atomic_coordinates,
  expanded_coords,
  unit_cell_metrics,
  cutoff_length = 7,
  x_diff_weight = 3,
  porous_adjustment = TRUE,
  distance_cutoffs = c(0.5, 1),
  cation_anion = FALSE,
  weighted_cn = FALSE
)
}
\arguments{
\item{distances}{Ignored (CrystalNN generates its own Voronoi basis).}

\item{atomic_coordinates}{Primary atom set.}

\item{expanded_coords}{Expanded supercell.}

\item{unit_cell_metrics}{Cell parameters.}

\item{cutoff_length}{Numeric. Cutoff in Angstroms for initial neighbor search.}

\item{x_diff_weight}{Numeric. Electronegativity difference weight.}

\item{porous_adjustment}{Logical. If TRUE, adjusts Voronoi weights.}

\item{distance_cutoffs}{Numeric vector. Penalizes neighbor distances greater than sum of radii.}

\item{cation_anion}{Logical. Restrictions targets to opposite charge.}

\item{weighted_cn}{Logical. Return fractional probabilities vs strict max probability.}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Rebuild of Pymatgen's \code{CrystalNN} algorithm from the ground up.
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{econ_nn}()},
\code{\link{minimum_distance}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
