% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crystract-package.R
\docType{package}
\name{crystract-package}
\alias{crystract}
\alias{crystract-package}
\title{crystract: Tools to Extract Data from CIF Files for Crystallography}
\description{
Parse Crystallographic Information Files (.cif) and extract essential
data such as chemical formulas, unit cell parameters, atomic coordinates,
and symmetry operations. Calculate interatomic distances, identify bonded
pairs using various algorithms (Minimum Distance, Brunner's, EconNN,
CrystalNN, Voronoi), determine nearest neighbor counts, and calculate bond
angles. Facilitates the preparation of crystallographic data for further
analysis, including machine learning applications in materials science.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://prabhulab.github.io/ml-crystals/}
}

}
\author{
\strong{Maintainer}: Don Ngo \email{dngo@carnegiescience.edu}

}
\keyword{internal}
