% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{econ_nn}
\alias{econ_nn}
\title{Identify Atomic Bonds using Hoppe's EconNN Method}
\usage{
econ_nn(distances, atomic_coordinates, tol = 0.2, use_fictive_radius = FALSE)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances.}

\item{atomic_coordinates}{A \code{data.table} of atomic coordinates used to map species to radii.}

\item{tol}{A bond strength cutoff (default 0.2).}

\item{use_fictive_radius}{Logical. If \code{TRUE}, calculates Hoppe's fictive ionic radius.}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Uses Effective Coordination Numbers (ECoN) and Mean Fictive Ionic Radii (MEFIR).
}
\examples{
dists <- data.table::data.table(Atom1 = c("Si1", "Si1"), Atom2 = c("O1", "O2"),
                                Distance = c(1.6, 2.0),
                                DeltaX = c(1, 0), DeltaY = c(0, 1), DeltaZ = c(0, 0))
ac <- data.table::data.table(Label = c("Si1", "O1", "O2"),
                             OxidationState = c(4, -2, -2))
econ_nn(dists, ac)
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{crystal_nn}()},
\code{\link{minimum_distance}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
