% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{econ_nn}
\alias{econ_nn}
\title{Identify Atomic Bonds using Hoppe's EconNN Method}
\usage{
econ_nn(distances, atomic_coordinates, tol = 0.2, cation_anion = FALSE)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances.}

\item{atomic_coordinates}{A \code{data.table} of atomic coordinates used to map species to radii.}

\item{tol}{A bond strength cutoff (default 0.2).}

\item{cation_anion}{Logical. If TRUE, restricts bonding to sites with opposite or zero charge.}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Uses Effective Coordination Numbers (ECoN) and Mean Fictive Ionic Radii (MEFIR).
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{crystal_nn}()},
\code{\link{minimum_distance}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
