% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cif_files.R
\name{export_analysis_to_csv}
\alias{export_analysis_to_csv}
\title{Export Analysis Results to a Directory of CSVs}
\usage{
export_analysis_to_csv(analysis_results, output_dir, overwrite = FALSE)
}
\arguments{
\item{analysis_results}{A \code{data.table} object.}

\item{output_dir}{Path to main output directory.}

\item{overwrite}{Logical.}
}
\value{
Invisibly returns output_dir.
}
\description{
Exports analysis results to CSV structure.
}
