% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{filter_by_elements}
\alias{filter_by_elements}
\title{Filter Distances by Element Symbols}
\usage{
filter_by_elements(distances, atomic_coordinates, elements_to_exclude)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances.}

\item{atomic_coordinates}{A \code{data.table} of atomic coordinates.}

\item{elements_to_exclude}{A character vector of element symbols to exclude.}
}
\value{
A \code{data.table} of distances with the specified elements removed.
}
\description{
Removes any distance pair where at least one of the atoms
corresponds to a specified list of element symbols.
}
\examples{
dists <- data.table::data.table(Atom1 = "Si1_1_0_0", Atom2 = "O1_1_0_0", Distance = 1.6)
ac <- data.table::data.table(Label = c("Si1", "O1"))
filter_by_elements(dists, ac, elements_to_exclude = "O")
}
\seealso{
Other property calculators: 
\code{\link{calculate_angles}()},
\code{\link{calculate_distances}()},
\code{\link{calculate_neighbor_counts}()},
\code{\link{calculate_weighted_neighbor_counts}()},
\code{\link{filter_atoms_by_symbol}()},
\code{\link{filter_by_wyckoff_symbol}()}
}
\concept{property calculators}
