% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_bonds.R
\name{minimum_distance}
\alias{minimum_distance}
\title{Identify Atomic Bonds using the Minimum Distance Method}
\usage{
minimum_distance(distances, delta = 0.1)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances from \code{calculate_distances}.}

\item{delta}{The relative tolerance parameter (default 0.1).}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Identifies bonded atoms by finding the nearest neighbor distance (d_min)
for each central atom and defining a cutoff distance (d_cut) as
d_cut = (1 + delta) * d_min.
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner_nn_reciprocal}()},
\code{\link{crystal_nn}()},
\code{\link{econ_nn}()},
\code{\link{voronoi_nn}()}
}
\concept{bonding algorithms}
