% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_properties.R
\name{minimum_distance}
\alias{minimum_distance}
\title{Identify Atomic Bonds using the Minimum Distance Method}
\usage{
minimum_distance(distances, delta = 0.1)
}
\arguments{
\item{distances}{A \code{data.table} of interatomic distances from \code{calculate_distances}.}

\item{delta}{The relative tolerance parameter (default 0.1 or ten percent).}
}
\value{
A \code{data.table} of bonded pairs.
}
\description{
Identifies bonded atoms by finding the nearest neighbor distance (d_min)
for each central atom and defining a cutoff distance (d_cut) as
d_cut = (1 + delta) * d_min.
}
\examples{
cif_file <- system.file("extdata", "ICSD422.cif", package = "crysmal")
if (file.exists(cif_file)) {
  # Setup code
  cif_content <- read_cif_files(cif_file)[[1]]
  atoms <- extract_atomic_coordinates(cif_content)
  metrics <- extract_unit_cell_metrics(cif_content)
  sym_ops <- extract_symmetry_operations(cif_content)
  full_cell <- apply_symmetry_operations(atoms, sym_ops)
  super_cell <- expand_transformed_coords(full_cell)
  dists <- calculate_distances(atoms, super_cell, metrics)

  # Apply Minimum Distance Method
  bonded <- minimum_distance(dists, delta = 0.1)
  print(head(bonded))
}
}
\seealso{
Other bonding algorithms: 
\code{\link{brunner}()},
\code{\link{hoppe}()}
}
\concept{bonding algorithms}
