% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cif_files.R
\name{read_cif_files}
\alias{read_cif_files}
\title{Read CIF Files into Memory}
\usage{
read_cif_files(file_paths)
}
\arguments{
\item{file_paths}{A character vector of paths to the CIF files.}
}
\value{
A list of \code{data.table} objects.
}
\description{
Reads one or more CIF files from disk and loads each into a
\code{data.table}. It includes encoding sanitization to handle legacy characters.
}
\examples{
cif_path <- system.file("extdata", "1590946.cif", package = "crystract")
if (file.exists(cif_path)) {
  cifs <- read_cif_files(cif_path)
}
}
