% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_cif_files.R
\name{read_cif_files}
\alias{read_cif_files}
\title{Read CIF Files into Memory}
\usage{
read_cif_files(file_paths)
}
\arguments{
\item{file_paths}{A character vector of paths to the CIF files.}
}
\value{
A list of \code{data.table} objects.
}
\description{
Reads one or more CIF files from disk and loads each into a
\code{data.table}.
}
\examples{
cif_file <- system.file("extdata", "ICSD422.cif", package = "crystract")
if (file.exists(cif_file)) {
  cif_data_list <- read_cif_files(cif_file)
  print(cif_data_list[[1]][1:5, ])
}
}
